refactor(board): rename translation files for clarity
board-translation.ts -> translation-cache.ts (board.strings is the
embedded cache the pure helpers read/apply), and translate-board.ts ->
resolve-translated-board.ts to match its sole export. Breaks the
near-mirror filenames and drops the folder-redundant prefixes.
